Create README.md
Hi there! I'm Bill

I am a Principal Tech Sourcer from Canada. I have helped great companies scale their technical teams and give Sourcing training to Recruitment teams. You are currently looking at a list of (recruitment) tools.

I created this because I wanted to centralize all the tools on 1 page for myself and others.

Hope this helps everyone that is in the same field! As this is on Github and open source, if you please this is something that interests you, please contribute.

All the tools are tested and testimony of all the hard work within my network of ensuring that the tools works and hope everyone shares the same success.

See this?